# spec/openapi.yaml
openapi: 3.0.0
info:
  title: PC Components Shop
  version: 0.0.0
tags: []
paths:
  /api/auth/logout:
    post:
      operationId: logout
      parameters: []
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/auth/me:
    get:
      operationId: me
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/auth/signin:
    post:
      operationId: signin
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SigninRequest'
  /api/auth/signup:
    post:
      operationId: signup
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
  /api/categories:
    get:
      operationId: listCategories
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryList'
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/orders:
    post:
      operationId: createOrder
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Ответ при отклонении заказа целиком с перечнем проблемных товаров.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/OrderInvalid'
                  - $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
    get:
      operationId: listMyOrders
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderList'
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/orders/{id}:
    get:
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/products:
    get:
      operationId: listProducts
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
          explode: false
        - name: search
          in: query
          required: false
          schema:
            type: string
          explode: false
        - name: priceMin
          in: query
          required: true
          schema:
            type: integer
            format: int32
            default: 0
          explode: false
        - name: priceMax
          in: query
          required: false
          schema:
            type: integer
            format: int32
          explode: false
        - name: available
          in: query
          required: false
          schema:
            type: boolean
          explode: false
        - name: page
          in: query
          required: true
          schema:
            type: integer
            format: int32
            default: 1
          explode: false
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            format: int32
            default: 12
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/products/{slug}:
    get:
      operationId: getProduct
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/promo:
    get:
      operationId: listPromoBlocks
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoBlockList'
        '400':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '401':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '409':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '422':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Единый формат ошибки API для всех ресурсов.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /health:
    get:
      operationId: health
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
components:
  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Машиночитаемый код ошибки.
        message:
          type: string
          description: Человекочитаемое сообщение для пользователя.
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
          description: Детализация по полям, если применимо.
      description: Единый формат ошибки API для всех ресурсов.
    Category:
      type: object
      required:
        - id
        - slug
        - name
      properties:
        id:
          type: integer
          format: int32
        slug:
          type: string
          description: Слаг для URL и фильтров.
        name:
          type: string
          description: Название категории.
      description: Категория товаров.
    CategoryList:
      type: object
      required:
        - categories
      properties:
        categories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
      description: Ответ со списком категорий.
    CreateOrderRequest:
      type: object
      required:
        - items
        - delivery
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemInput'
        delivery:
          $ref: '#/components/schemas/DeliveryDetails'
      description: Тело запроса на оформление заказа. Итог считает сервер по текущим ценам.
    DeliveryDetails:
      type: object
      required:
        - method
        - recipientName
        - phone
      properties:
        method:
          $ref: '#/components/schemas/DeliveryMethod'
        recipientName:
          type: string
          description: Имя получателя.
        phone:
          type: string
          description: Телефон получателя.
        address:
          type: string
          description: Адрес одной строкой — только для доставки.
      description: Данные получения. Адрес обязателен только при доставке (проверяет сервер).
    DeliveryMethod:
      type: string
      enum:
        - delivery
        - pickup
      description: Способ получения. Город подразумевается один.
    ErrorDetail:
      type: object
      required:
        - message
      properties:
        field:
          type: string
          description: Поле, к которому относится ошибка (может отсутствовать).
        message:
          type: string
          description: Человекочитаемое описание ошибки.
      description: Детализированная ошибка по полю.
    HealthStatus:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
      description: Проверка работоспособности сервиса.
    Order:
      type: object
      required:
        - id
        - delivery
        - items
        - total
        - status
        - createdAt
      properties:
        id:
          type: integer
          format: int32
        delivery:
          $ref: '#/components/schemas/DeliveryDetails'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemSnapshot'
        total:
          allOf:
            - $ref: '#/components/schemas/rubles'
          description: Итоговая сумма в рублях, посчитанная сервером.
        status:
          type: string
          enum:
            - paid
        createdAt:
          type: string
          format: date-time
      description: Заказ. Оплаты нет — статус сразу 'Оплачен' и не меняется.
    OrderInvalid:
      type: object
      required:
        - code
        - message
        - items
      properties:
        code:
          type: string
          enum:
            - ORDER_INVALID
        message:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemProblem'
      description: Ответ при отклонении заказа целиком с перечнем проблемных товаров.
    OrderItemInput:
      type: object
      required:
        - productId
        - quantity
      properties:
        productId:
          type: integer
          format: int32
        quantity:
          type: integer
          format: int32
          minimum: 1
      description: Позиция корзины в запросе на оформление. Цена не передаётся.
    OrderItemProblem:
      type: object
      required:
        - productId
        - code
      properties:
        productId:
          type: integer
          format: int32
        code:
          type: string
          enum:
            - NOT_FOUND
            - UNAVAILABLE
        name:
          type: string
      description: Проблемная позиция при атомарном отказе.
    OrderItemSnapshot:
      type: object
      required:
        - productId
        - name
        - price
        - quantity
      properties:
        productId:
          type: integer
          format: int32
        name:
          type: string
        price:
          $ref: '#/components/schemas/rubles'
        quantity:
          type: integer
          format: int32
      description: Снимок товара в позиции заказа — название и цена на момент покупки.
    OrderList:
      type: object
      required:
        - orders
      properties:
        orders:
          type: array
          items:
            $ref: '#/components/schemas/Order'
      description: Список заказов пользователя.
    Pagination:
      type: object
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int32
          description: Общее количество товаров по фильтру.
        totalPages:
          type: integer
          format: int32
          description: Количество страниц.
      description: Метаданные пагинации.
    Product:
      type: object
      required:
        - id
        - slug
        - name
        - description
        - price
        - available
        - category
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int32
        slug:
          type: string
          description: Короткий слаг для URL.
        name:
          type: string
          description: Название товара.
        description:
          type: string
          description: Краткое описание.
        price:
          allOf:
            - $ref: '#/components/schemas/rubles'
          description: Цена в рублях — целое число без копеек.
        imageUrl:
          type: string
          description: URL изображения. Если отсутствует — показать заглушку.
        available:
          type: boolean
          description: Доступен ли товар к заказу.
        category:
          allOf:
            - $ref: '#/components/schemas/Category'
          description: Категория — ровно одна, без вложенности.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      description: Товар каталога.
    ProductList:
      type: object
      required:
        - products
        - pagination
      properties:
        products:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        pagination:
          allOf:
            - $ref: '#/components/schemas/Pagination'
          description: Метаданные пагинации.
      description: Страница каталога с метаданными пагинации.
    PromoBlock:
      type: object
      required:
        - id
        - title
        - text
        - product
      properties:
        id:
          type: integer
          format: int32
        title:
          type: string
          description: Заголовок блока.
        text:
          type: string
          description: Текст блока.
        product:
          allOf:
            - $ref: '#/components/schemas/Product'
          description: Рекламируемый товар — целиком внутри блока.
      description: 'Промо-блок главной страницы: реклама конкретного товара.'
    PromoBlockList:
      type: object
      required:
        - promoBlocks
      properties:
        promoBlocks:
          type: array
          items:
            $ref: '#/components/schemas/PromoBlock'
      description: Ответ со списком промо-блоков.
    SigninRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email — логин покупателя.
        password:
          type: string
          description: Пароль.
      description: Тело запроса входа.
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: Email — логин покупателя. Если адрес уже занят, сервер вернёт 409 EMAIL_TAKEN.
        password:
          type: string
          minLength: 8
          description: Пароль. Минимум 8 символов.
      description: Тело запроса регистрации.
    User:
      type: object
      required:
        - id
        - email
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int32
          description: Уникальный идентификатор пользователя.
        email:
          type: string
          format: email
          description: Email — логин пользователя.
        createdAt:
          type: string
          format: date-time
          description: Дата создания учётной записи.
        updatedAt:
          type: string
          format: date-time
          description: Дата последнего обновления.
      description: "Пользователь магазина: учётная запись и покупатель — одно и то же, отдельной сущности 'клиент' в системе нет."
    rubles:
      type: integer
      format: int32
      minimum: 0
      description: Сумма в рублях — целое число без копеек.
servers:
  - url: http://localhost:3000
    description: Local development server
    variables: {}